requirements.txt
requests>=2.26.0
python-dotenv>=1.0.0
cryptography>=41.0.0

